+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsFastInitBoxTestGenerated.java
@@ -2442,9 +2442,21 @@ public class JsFastInitBoxTestGenerated extends AbstractJsFastInitBoxTest {
       }
 
       @Test
+      @TestMetadata("GenericMemberInjectionAcrossModules.kt")
+      public void testGenericMemberInjectionAcrossModules() {
+        run("GenericMemberInjectionAcrossModules.kt");
+      }
+
+      @Test
       @TestMetadata("GenericMemberInjectionFromGraph.kt")
       public void testGenericMemberInjectionFromGraph() {
         run("GenericMemberInjectionFromGraph.kt");
+      }
+
+      @Test
+      @TestMetadata("GenericMemberInjectionSameModule.kt")
+      public void testGenericMemberInjectionSameModule() {
+        run("GenericMemberInjectionSameModule.kt");
       }
 
       @Test